Enabling puppetdb API v4
This commit allows the user to specify the API v4 in the URL.

There are no changes in the syntax query in the puppetdb API v4, even
though such version is still experimental. So the PuppetdbConnector
logic does not change: it is still up to the client code to specify an
endpoint compatible with the requested API version.

Adding changelog file for possible release.